test(input): Rename InputViewTest.kt to InputParserTest.kt
